fix STARTS_WITH function
